+++ packages/Phased/State/Exceptions/VuexMissingRequiredParameter.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Phased\State\Exceptions;
+
+use Exception;
+
+class VuexMissingRequiredParameter extends Exception {
+
+}